+++ src/api/errors/index.ts
@@ -3,4 +3,5 @@ export * from './api';
 export * from './chain';
 export * from './election';
 export * from './file';
+export * from './faucet';
 export * from './vote';