+++ test/api/vote.test.ts
@@ -1,7 +1,15 @@
-import { Election, EnvOptions, PlainCensus, VocdoniSDKClient, Vote } from '../../src';
+import {
+  Election,
+  ElectionStatus,
+  EnvOptions,
+  ErrElectionFinished,
+  ErrElectionNotStarted,
+  PlainCensus,
+  VocdoniSDKClient,
+  Vote,
+} from '../../src';
 // @ts-ignore
 import { URL, setFaucetURL } from './util/client.params';
-import { ErrElectionNotStarted } from '../../src';
 import { Wallet } from '@ethersproject/wallet';
 
 let client: VocdoniSDKClient;
@@ -21,8 +29,7 @@ const createElection = (census, electionType?, voteType?) => {
   const election = Election.from({
     title: 'SDK Testing - Title',
     description: 'SDK Testing - Description',
-    startDate: new Date().getTime() + 10000000,
-    endDate: new Date().getTime() + 20000000,
+    endDate: new Date().getTime() + 12000,
     census,
     electionType: electionType ?? null,
     voteType: voteType ?? null,
@@ -43,7 +50,7 @@ const createElection = (census, electionType?, voteType?) => {
 };
 
 describe('Vote API tests', () => {
-  it('should throw trying to vote when election has not started', async () => {
+  it('should throw trying to vote when election has not started and when is already finished', async () => {
     const voter = Wallet.createRandom();
     const census = new PlainCensus();
     census.add(await voter.getAddress());
@@ -59,5 +66,14 @@ describe('Vote API tests', () => {
     await expect(async () => {
       await client.submitVote(vote);
     }).rejects.toThrow(ErrElectionNotStarted);
+
+    let publishedElection;
+    do {
+      publishedElection = await client.fetchElection(electionId);
+    } while (publishedElection.status !== ElectionStatus.ENDED);
+
+    await expect(async () => {
+      await client.submitVote(vote);
+    }).rejects.toThrow(ErrElectionFinished);
   }, 85000);
 });